# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: User Management Control Panel API
  version: 0.1.0
  description: |
    Control-plane API for the user-management + licensing system.

    Covers admin authentication, organization/user/RBAC management, plan & subscription
    administration, license JWKS publication, key rotation, usage ingestion, audit log
    access, SSO configuration, API key management, and the domain event stream.

    Two authentication schemes are supported:
      - **bearerAuth**: short-lived admin session JWT, issued by `POST /api/v1/auth/login`.
      - **apiKeyAuth**: long-lived programmatic key, sent as `Authorization: ApiKey <token>`.

    The public `/.well-known/jwks.json` endpoint is unauthenticated.

    Error responses follow [RFC 7807 Problem Details](https://datatracker.ietf.org/doc/html/rfc7807).

    > **Caveat:** the license-issuance endpoint (`POST .../licenses`) referenced in the
    > implementation plan is *not yet wired up* in `control-panel-api`. It is documented
    > here as the intended contract for Phase 1 closure.
  contact:
    name: Control Panel Team
    email: support@control-panel.example.com
  license:
    name: Proprietary
servers:
  - url: https://control-panel.example.com
    description: Production
  - url: https://staging.control-panel.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Exchange email + password for an admin session JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/password-reset/request:
    post:
      tags: [auth]
      summary: Trigger a password-reset email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Always 200, regardless of whether the email exists (no enumeration).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  reset_token:
                    type: string
                    description: Only present when `app.auth.expose-reset-token=true` (dev/test).
  /api/v1/auth/password-reset/confirm:
    post:
      tags: [auth]
      summary: Complete a password reset using a token from the email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '204':
          description: Password updated.
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Return the currently authenticated user, org memberships, and granted permissions.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # orgs
  # ============================================================
...
  /api/v1/audit:
    get:
      tags: [audit]
      summary: Search the global audit log.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: actor
          schema: { type: string, format: uuid }
        - in: query
          name: targetType
          schema: { type: string }
        - in: query
          name: targetId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Paged audit entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEntryPage'
  /api/v1/orgs/{orgId}/audit:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [audit]
      summary: Search audit entries scoped to a single organization.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: actor
          schema: { type: string, format: uuid }
        - in: query
          name: targetType
          schema: { type: string }
        - in: query
          name: targetId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Paged audit entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEntryPage'

  # ============================================================
  # sso
  # ============================================================
...
  /api/v1/orgs/{orgId}/api-keys:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [apikeys]
      summary: List API keys for an org.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
    post:
      tags: [apikeys]
      summary: Create an org-scoped API key. The plaintext key is returned ONCE.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '200':
          description: Created. Save the `key` value — it is shown only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
  /api/v1/orgs/{orgId}/api-keys/{apiKeyId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - in: path
        name: apiKeyId
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [apikeys]
      summary: Revoke an API key.
      responses:
        '204':
          description: Revoked.

  # ============================================================
  # events
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session JWT from `POST /api/v1/auth/login`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Long-lived programmatic key.
        Format: `Authorization: ApiKey <token>` where `<token>` is the value returned
        once by `POST /api/v1/orgs/{orgId}/api-keys`.
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string, format: uuid }
...
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Request body or parameters invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Bad Request' }
        status: { type: integer, example: 400 }
        detail: { type: string, example: 'Field `email` is required.' }
        instance: { type: string, format: uri }
        code: { type: string, example: 'validation_failed' }
      required: [title, status]
...
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
      required: [total, page, size]

    # ---- auth ----
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken: { type: string, description: 'Session JWT.' }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/User' }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirm:
      type: object
      required: [token, newPassword]
      properties:
        token: { type: string }
        newPassword: { type: string, format: password, minLength: 8 }
    MeResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        orgs:
          type: array
          items: { $ref: '#/components/schemas/OrgMembership' }
        permissions:
          type: array
          items: { type: string }
      required: [user, orgs, permissions]

    # ---- user ----
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status:
          type: string
          enum: [ACTIVE, INVITED, DISABLED, DELETED]
        superAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
      required: [id, email, status]
...
    OrgMembership:
      type: object
      properties:
        orgId: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER, VIEWER]
      required: [orgId, role]
...
    AuditEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        actorUserId: { type: string, format: uuid, nullable: true }
        actorOrgId: { type: string, format: uuid, nullable: true }
        action: { type: string, example: 'subscription.created' }
        targetType: { type: string }
        targetId: { type: string }
        payloadJson: { type: string }
        ipAddress: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }
      required: [id, action, occurredAt]
    AuditEntryPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AuditEntry' }
          required: [items]

    # ---- sso ----
...
    ApiKey:
      type: object
      properties:
        id: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        name: { type: string }
        keyPrefix: { type: string }
        scopes: { type: string, description: 'JSON-encoded list of scope codes.' }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
      required: [id, orgId, name, keyPrefix, createdAt]
    ApiKeyCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
        scopes:
          type: array
          items: { type: string }
    ApiKeyCreateResponse:
      type: object
      required: [id, name, key, keyPrefix, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        key:
          type: string
          description: |
            Plaintext token. Shown only at creation; the server stores only a hashed copy.
        keyPrefix: { type: string }
        scopes:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }

    # ---- events ----
...